+++ .env
@@ -1,0 +1,1 @@
+KONTENT_PROJECT_ID=38ce511d-359d-008e-7529-50e7f0cb4b82
