+++ package.json
@@ -17,5 +17,10 @@
   },
   "resolutions": {
     "**/**/axios": "^0.21.1"
+  },
+  "devDependencies": {
+    "@types/node": "^18.7.6",
+    "@types/react": "^18.0.17",
+    "typescript": "^4.7.4"
   }
 }
